UDP updates UDP connectivity added to plugin.
Moved json functions into utility.
