+++ src/DataProcessor.h
@@ -13,31 +13,6 @@
 using namespace std;
 
 
-
-struct dataStruct {
-    string dref;
-    string description;
-    int index;
-    string units;
-    int freq;
-    int rounding;
-    string value;
-
-};
-
-struct dataFrame{
-    vector<dataStruct> state;
-    vector<dataStruct> inputs;
-    vector<dataStruct> instructions;
-    vector<dataStruct> failures;
-};
-
-struct dataReference {
-    XPLMDataRef dataref;
-    string type;
-};
-
-
 class DataProcessor
 {
     public:
